Fix a race that can occur if a subprocess forks and issues a syscall
It seems like, in this case (which is not very rare), the forked process issues
a system call (likely an exec) before the event processing loop can process the
pseudo-STOP event after the attach.  This causes the "impossible" case to
trigger.  To fix that, just assume that if we are in this case that we are
entering a syscall stop (and not exiting a syscall stop).  This should be the
case, because we won't actually miss any real events (so syscall enter must
precede syscall exit).