Handle out-of-order attach-stops
An attach-stop can be delivered to the tracer before the ptrace-event-stop
where we "should" first learn about the new tracee. Handle this by marking the
new tracee when we see a `SIGSTOP` for an untracked pid, and making sure we
don't reset the state to `Attaching` in the out-of-order ptrace-event-stop.