Use `std::process::Command` with `Ptracer::spawn()` (#21)
Accept a `std::process::Command` in `Ptracer::spawn()`, and return a `Child` instead of a `Tracee`. This makes the ctor do much less work, and in particular, prevents hanging on `wait()`. Move the existing `set_options()` call into our interpretation of `SIGTRAP` stops. This uses the new `Spawning` state to help interpret the legacy `SIGTRAP` issued for the `execve(2)` from the self-attached tracee.

Closes #11.