Remove redundant PID fields (#50)
Almost all `Stop` variants have a `Pid`-valued field that identifies the stopped tracee. This is redundant when accessed via the `Tracee`, which we typically expect any `Stop` consumer to have access to. This is also confusing, since we use tuple struct variants, and there are sometimes multiple PIDs associated with a stop.

Remove the redundant PIDs. Require users to pass either the `Tracee` or tracee PID when needed.

This is a breaking change.